+++ ci/tests/e2e-ansible.sh
@@ -54,8 +54,33 @@ test_operator() {
     if ! timeout 1m kubectl rollout status deployment/memcached-operator;
     then
         echo FAIL: operator failed to run
+        kubectl describe pods
         kubectl logs deployment/memcached-operator -c operator
         kubectl logs deployment/memcached-operator -c ansible
+        exit 1
+    fi
+
+    # verify that metrics service was created
+    if ! timeout 20s bash -c -- "until kubectl get service/memcached-operator-metrics > /dev/null 2>&1; do sleep 1; done";
+    then
+        echo "Failed to get metrics service"
+        kubectl logs deployment/memcached-operator
+        exit 1
+    fi
+
+    # verify that the metrics endpoint exists
+    if ! timeout 1m bash -c -- "until kubectl run -it --rm --restart=Never test-metrics --image=registry.access.redhat.com/ubi7/ubi-minimal:latest -- curl -sfo /dev/null http://memcached-operator-metrics:8383/metrics; do sleep 1; done";
+    then
+        echo "Failed to verify that metrics endpoint exists"
+        kubectl logs deployment/memcached-operator
+        exit 1
+    fi
+
+    # verify that the operator metrics endpoint exists
+    if ! timeout 1m bash -c -- "until kubectl run -it --rm --restart=Never test-metrics --image=registry.access.redhat.com/ubi7/ubi-minimal:latest -- curl -sfo /dev/null http://memcached-operator-metrics:8686/metrics; do sleep 1; done";
+    then
+        echo "Failed to verify that metrics endpoint exists"
+        kubectl logs deployment/memcached-operator
         exit 1
     fi
 
@@ -64,14 +89,25 @@ test_operator() {
     if ! timeout 20s bash -c -- 'until kubectl get deployment -l app=memcached | grep memcached; do sleep 1; done';
     then
         echo FAIL: operator failed to create memcached Deployment
+        kubectl describe pods
         kubectl logs deployment/memcached-operator -c operator
         kubectl logs deployment/memcached-operator -c ansible
         exit 1
     fi
+
+    # verify that metrics reflect cr creation
+    if ! bash -c -- 'kubectl run -it --rm --restart=Never test-metrics --image=registry.access.redhat.com/ubi7/ubi-minimal:latest -- curl http://memcached-operator-metrics:8686/metrics | grep example-memcached';
+    then
+        echo "Failed to verify custom resource metrics"
+        kubectl logs deployment/memcached-operator
+        exit 1
+    fi
+
     memcached_deployment=$(kubectl get deployment -l app=memcached -o jsonpath="{..metadata.name}")
     if ! timeout 1m kubectl rollout status deployment/${memcached_deployment};
     then
         echo FAIL: memcached Deployment failed rollout
+        kubectl describe pods
         kubectl logs deployment/${memcached_deployment}
         exit 1
     fi
@@ -86,6 +122,7 @@ test_operator() {
     if kubectl get configmap deleteme 2> /dev/null;
     then
         echo FAIL: the finalizer did not delete the configmap
+        kubectl describe pods
         kubectl logs deployment/memcached-operator -c operator
         kubectl logs deployment/memcached-operator -c ansible
         exit 1
@@ -95,6 +132,7 @@ test_operator() {
     if ! timeout 20s bash -c -- "while kubectl get deployment ${memcached_deployment} 2> /dev/null; do sleep 1; done";
     then
         echo FAIL: memcached Deployment did not get garbage collected
+        kubectl describe pods
         kubectl logs deployment/memcached-operator -c operator
         kubectl logs deployment/memcached-operator -c ansible
         exit 1
@@ -104,6 +142,7 @@ test_operator() {
     if kubectl logs deployment/memcached-operator -c operator | grep -i error;
     then
         echo FAIL: the operator log includes errors
+        kubectl describe pods
         kubectl logs deployment/memcached-operator -c operator
         kubectl logs deployment/memcached-operator -c ansible
         exit 1
@@ -135,6 +174,7 @@ remove_operator
 if ! timeout 60s bash -c -- "until kubectl get pods -l name=memcached-operator |& grep \"No resources found\"; do sleep 2; done";
 then
     echo FAIL: memcached-operator Deployment did not get garbage collected
+    kubectl describe pods
     kubectl logs deployment/memcached-operator -c operator
     kubectl logs deployment/memcached-operator -c ansible
     exit 1