+++ hack/tests/scorecard-subcommand.sh
@@ -25,6 +25,7 @@ echo $commandoutput | grep "Total Score: 82%"
 # test config file
 commandoutput2="$(operator-sdk scorecard \
   --proxy-image "$DEST_IMAGE" \
+  --proxy-pull-policy "Never" \
   --config "$CONFIG_PATH")"
 # check basic suite
 echo $commandoutput2 | grep '^.*"error": 0,[[:space:]]"pass": 3,[[:space:]]"partialPass": 0,[[:space:]]"fail": 0,[[:space:]]"totalTests": 3,[[:space:]]"totalScorePercent": 100,.*$'