+++ internal/pkg/scaffold/ansible/gopkgtoml.go
@@ -19,7 +19,6 @@ import (
 
 	"github.com/operator-framework/operator-sdk/internal/pkg/scaffold"
 	"github.com/operator-framework/operator-sdk/internal/pkg/scaffold/input"
-	"github.com/operator-framework/operator-sdk/internal/pkg/scaffold/internal/deps"
 )
 
 // GopkgToml - the Gopkg.toml file for a hybrid operator
@@ -38,8 +37,8 @@ func (s *GopkgToml) GetInput() (input.Input, error) {
 const gopkgTomlTmpl = `[[constraint]]
   name = "github.com/operator-framework/operator-sdk"
   # The version rule is used for a specific release and the master branch for in between releases.
-  # branch = "master" #osdk_branch_annotation
-  version = "=v0.9.0" #osdk_version_annotation
+  branch = "master" #osdk_branch_annotation
+  # version = "=v0.10.0" #osdk_version_annotation
 
 [[override]]
   name = "k8s.io/api"
@@ -58,10 +57,7 @@ const gopkgTomlTmpl = `[[constraint]]
   unused-packages = true
 `
 
-func PrintDepGopkgTOML(asFile bool) error {
-	if asFile {
-		_, err := fmt.Println(gopkgTomlTmpl)
-		return err
-	}
-	return deps.PrintDepGopkgTOML(gopkgTomlTmpl)
+func PrintDepGopkgTOML() error {
+	_, err := fmt.Println(gopkgTomlTmpl)
+	return err
 }