+++ internal/pkg/scaffold/gopkgtoml.go
@@ -18,7 +18,6 @@ import (
 	"fmt"
 
 	"github.com/operator-framework/operator-sdk/internal/pkg/scaffold/input"
-	"github.com/operator-framework/operator-sdk/internal/pkg/scaffold/internal/deps"
 )
 
 const GopkgTomlFile = "Gopkg.toml"
@@ -83,8 +82,8 @@ required = [
 [[constraint]]
   name = "github.com/operator-framework/operator-sdk"
   # The version rule is used for a specific release and the master branch for in between releases.
-  # branch = "master" #osdk_branch_annotation
-  version = "=v0.9.0" #osdk_version_annotation
+  branch = "master" #osdk_branch_annotation
+  # version = "=v0.10.0" #osdk_version_annotation
 
 [prune]
   go-tests = true
@@ -93,13 +92,9 @@ required = [
   [[prune.project]]
     name = "k8s.io/kube-state-metrics"
     unused-packages = true
-
 `
 
-func PrintDepGopkgTOML(asFile bool) error {
-	if asFile {
-		_, err := fmt.Println(gopkgTomlTmpl)
-		return err
-	}
-	return deps.PrintDepGopkgTOML(gopkgTomlTmpl)
+func PrintDepGopkgTOML() error {
+	_, err := fmt.Println(gopkgTomlTmpl)
+	return err
 }